refactor(frontend): extract TrackerTab (15/N)
Application tracker + resume version manager + compare UI.
~310 LOC, the largest tab so far. ~30 props passed via state object.

App.jsx 2483 → 2175 LOC. Build clean.
